Ryddet litt opp i små feil og endret litt estetik
