+++ Kartverket/Views/Home/AreaChangeOverview.cshtml
@@ -1,4 +1,4 @@
-﻿@model List<Kartverket.Models.GeoChange>
+﻿@model List<Kartverket.Data.GeoChange>
 @*@model List<Kartverket.Models.AreaChange>*@
 
 @*
